% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_reads.R
\name{standart_normalize}
\alias{standart_normalize}
\title{standart_normalize}
\usage{
standart_normalize(
  fq.to.normalize,
  count.reads = FALSE,
  number.replicates,
  sample.reads,
  parallel.core = parallel::detectCores() - 1
)
}
\description{
main normalize function
}
\keyword{internal}
